Rename module to ground_texture_sim
