--- data_generation/__init__.py
@@ -1,8 +1,0 @@
-"""!
-The data_generation module contains all the logic code to load information, manipulate the
-simulation, and write the data. It also contains the unit tests for each function.
-"""
-from data_generation import blender_interface
-from data_generation import configuration_loader
-from data_generation import data_output
-from data_generation import name_configuration
