+++ generate_data.py
@@ -9,7 +9,7 @@ specified directory.
 SPDX-License-Identifier: GPL-3.0-or-later
 """
 import sys
-import data_generation
+import ground_texture_sim
 
 
 def main() -> None:  # pragma: no cover
@@ -21,16 +21,17 @@ def main() -> None:  # pragma: no cover
 
     @return None
     """
-    config_dict, trajectory_list = data_generation.configuration_loader.load_configuration(
+    config_dict, trajectory_list = ground_texture_sim.configuration_loader.load_configuration(
         sys.argv)
-    data_generation.blender_interface.generate_images(
+    ground_texture_sim.blender_interface.generate_images(
         config_dict, trajectory_list)
-    data_generation.data_output.prepare_output_folder(config_dict['output'])
-    data_generation.data_output.write_camera_intrinsic_matrix(
+    ground_texture_sim.data_output.prepare_output_folder(config_dict['output'])
+    ground_texture_sim.data_output.write_camera_intrinsic_matrix(
         config_dict['camera']['name'], config_dict['output'])
-    data_generation.data_output.write_camera_pose(
+    ground_texture_sim.data_output.write_camera_pose(
         config_dict['camera'], config_dict['output'])
-    data_generation.data_output.write_list_files(config_dict, trajectory_list)
+    ground_texture_sim.data_output.write_list_files(
+        config_dict, trajectory_list)
 
 
 if __name__ == '__main__':  # pragma: no cover
