+++ ground_texture_sim/__init__.py
@@ -1,0 +1,8 @@
+"""!
+The data_generation module contains all the logic code to load information, manipulate the
+simulation, and write the data. It also contains the unit tests for each function.
+"""
+from ground_texture_sim import blender_interface
+from ground_texture_sim import configuration_loader
+from ground_texture_sim import data_output
+from ground_texture_sim import name_configuration
