+++ ground_texture_sim/name_configuration.py
